requirements.txt
# Core ML & Deep Learning
tensorflow>=2.14.0,<3.0
scikit-learn>=1.3.0
numpy>=1.24.0
pandas>=2.0.0
opencv-python>=4.8.0

# Web Framework
Flask>=2.3.0
Flask-CORS>=4.0.0
Flask-RESTful>=0.3.10

# Database
SQLAlchemy>=2.0.0
python-dotenv>=1.0.0

# Image processing
Pillow>=10.0.0
matplotlib>=3.7.0

# Weather & data
requests>=2.31.0

# Utilities
PyJWT>=2.8.0
bcrypt>=4.0.1
python-dateutil>=2.8.0
pytz>=2023.0
